+++ src/libc/ansi/stdio/flsbuf.c
@@ -1,3 +1,4 @@
+/* Copyright (C) 2003 DJ Delorie, see COPYING.DJ for details */
 /* Copyright (C) 2001 DJ Delorie, see COPYING.DJ for details */
 /* Copyright (C) 1999 DJ Delorie, see COPYING.DJ for details */
 /* Copyright (C) 1996 DJ Delorie, see COPYING.DJ for details */
@@ -17,7 +18,7 @@ int
 _flsbuf(int c, FILE *f)
 {
   char *base;
-  int n;
+  ssize_t n;
   size_t rn;
   char c1;
   size_t size;
